eliminate old button test, port to new button test
